--- utils/config.go
@@ -1,29 +1,0 @@
-package utils
-
-import "github.com/spf13/viper"
-
-type Config struct {
-	DBDriver      string `mapstructure:"DB_DRIVER"`
-	DBSource      string `mapstructure:"DB_SOURCE"`
-	ServerAddress string `mapstructure:"SERVER_ADDRESS"`
-}
-
-func LoadConfig(path string) (config Config, err error) {
-	// ディレクトリパス
-	viper.AddConfigPath(path)
-	// 拡張子の前
-	viper.SetConfigName("app")
-	// 拡張子
-	viper.SetConfigType("env")
-
-	viper.AutomaticEnv()
-
-	err = viper.ReadInConfig()
-
-	if err != nil {
-		return
-	}
-
-	err = viper.Unmarshal(&config)
-	return
-}
